Setup subdomain routing
When `BASE_URL` has been set, rewrite request to prefix the pathname
with the provided subdomain. A subdomain must always be provided -
assumes that the root domain (example.com) is redirected (eg. to
www.example.com).

When `BASE_URL` has not been set, redirects `/` to `/www` for an easier
time testing. This will likely apply in local and preview environments.